Terminados todos los puntos, falta seguridad y revisar error handling

+++ src/main/java/com/parcial/pruebaTp/dtos/TarifaDto.java
@@ -11,10 +11,10 @@ public class TarifaDto {
     private long id;
     private long tipoTarifa;
     private String definicion;
-    private long diaSemana;
-    private long diaMes;
-    private long mes;
-    private long anio;
+    private Integer diaSemana;
+    private Integer diaMes;
+    private Integer mes;
+    private Integer anio;
     private double montoFijoAlquiler;
     private double montoMinutoFraccion;
     private double montoKm;
